+++ CHANGES.rst
@@ -5,7 +5,7 @@ dev
 ---
 
 * [feature] ``export`` command to change file format.  Useful for keys.
-* [feature] ``export-pubkey`` command to extract public key
+* [feature] ``export-pub`` command to extract public key
   from certificate or private key.
 * [feature] Support DER output format.
 * [cleanup] Split code into submodules.